rename: lattice files
and fix imports
