+++ src/python/impactx/dashboard/Input/lattice/utils.py
