+++ src/python/impactx/dashboard/Run/simulation.py
@@ -8,7 +8,7 @@ License: BSD-3-Clause-LBNL
 
 from .. import state
 from ..Input.distribution.utils import DistributionFunctions
-from ..Input.latticeConfiguration.latticeMain import parameter_input_checker_for_lattice
+from ..Input.lattice.ui import parameter_input_checker_for_lattice
 
 TRACKING_MODE_COMMANDS = {
     "Particle Tracking": """\
